cr fix 2 - ALPageField
